school_proj/requirements.txt
asgiref==3.8.1
djando==0.2
Django==4.2.21
djangorestframework==3.16.0
psycopg==3.2.8
psycopg-binary==3.2.8
psycopg2==2.9.1
Pygments==2.9.0
pytz==2021.1
redgreenunittest==0.1.1
sqlparse==0.4.1
typing_extensions==4.13.2
